Added support for deleting and editing objects using fillout form.
The form field spec is specified in the method form_fields(), which
is empty in FS but should be overwritten by the child object.

Two methods (and modes) were added: edit_object() and delete_object().
They were taken from what we had in Account/Pay Method and were
generalized.

Andrew, I hope the way this is done here is OK.  Let me know if you
have a better idea. :)

-marcos