Version 1.61, fixed a get_cookie() bug
In a situation that there was more than one call in a web session to
add_cookie() with the same cookie name, but possibly different path and
domain the code used to return the first cookie value. Now it returns
the last.

Both are not ideal, but the use case for this is so rare that coding
actual checks for matching domain and path is not worth it. Returning
the last set value is a compromise allowing most real usages to work.