1.73 pageurl() returns real unmodified URL
This was an old problem where for URLs like http://sample.com/foo/
a call to pageurl() would return http://sample.com/foo/index.html
(appending index.html). This affected all of the derived code, most
importantly <%URL...%> template calls.

Changed to return the actual URL as was requested, not as it was
interpreted in file templates.

Also fixed benchmark stats sort stability in testing for
https://rt.cpan.org/Public/Bug/Display.html?id=127440